生成パラメータのリセットボタンを追加、スマホの場合、生成ボタン、ステータスを画面下に固定する、Seedの説明文を改善する
#257 #258 #259
