+++ app/[lang]/generation/_hooks/use-image-generation-machine.tsx
@@ -55,8 +55,9 @@ export const useImageGenerationMachine = (props: Props) => {
     send({ type: "UPDATE_CONFIG", value })
   }
 
-  const init = () => {
-    const value = action.init().getState()
+  const reset = () => {
+    cacheStorage.reset()
+    const value = action.reset().getState()
     send({ type: "UPDATE_CONFIG", value })
   }
 
@@ -192,7 +193,7 @@ export const useImageGenerationMachine = (props: Props) => {
 
   return {
     context: new ImageGenerationContextView(state.context),
-    init,
+    reset,
     updateSettings,
     updateModelId,
     changeLoraModel: changeLoraConfig,
